# libc/libfread: Use memcpy to copy read-ahead buffer to caller buffer.